requirements.txt
aiogram==3.10.0
httpx==0.26.0
python-dotenv==1.0.0
google-api-python-client==2.120.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
cryptography==42.0.5
python-magic==0.4.27
